General cleanup for the compiler.
Using gcc 9.2.1, and adding in "-Wextra" found a bunch of things
that needed cleaning up:
 * many unused parameters
 * a bunch of signed-vs-unsigned comparisons
 * a few "fallthrough" cases not correctly marked
 * a struct array partion initialization that didn't always use tags
 * renamed usage() to usage_and_exit() to be clear
 * some duplicate struct initialzation tags/values
 * a couple of funky "else if" formats
 * make SOCK_DEBUG_VERBOSE actually work with a fwd declaration
 * fixed a couple of switch statements without a default case

 Should be no functional change.

+++ message.c
@@ -325,7 +325,7 @@ __revert_fqdn(const char *prefix, const char *__fqdn, const char *suffix)
 		if (*comp == '\0')
 			continue;
 		comp_len = strlen(comp);
-		if (pos + comp_len + 2 > sizeof(namebuf)) {
+		if (pos + comp_len + 2 > (int)sizeof(namebuf)) {
 			isns_error("%s: FQDN too long\n", __FUNCTION__);
 			goto out;
 		}
@@ -341,7 +341,7 @@ __revert_fqdn(const char *prefix, const char *__fqdn, const char *suffix)
 	if (suffix) {
 		int	sfx_len = strlen(suffix);
 
-		if (pos + sfx_len + 2 > sizeof(namebuf)) {
+		if (pos + sfx_len + 2 > (int)sizeof(namebuf)) {
 			isns_error("%s: name too long\n", __FUNCTION__);
 			goto out;
 		}
